requirements.txt
anyio==3.7.1
click==8.1.7
fastapi==0.105.0
h11==0.14.0
httptools==0.6.1
idna==3.6
marshmallow==3.20.1
packaging==24.1
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
openai==1.3.9 
colorlog==6.7.0
ruff==0.5.5
isort==5.13.2
tox==4.21.2
polygon-api-client==1.14.2
